Release  0.3.4

See merge request mr/python_base_app!13

+++ python_base_app/daemon.py
@@ -25,14 +25,15 @@ Changes:        23rd Jan 2009 (David Mytton <david@boxedice.com>)
 # Core modules
 from __future__ import print_function
 
-from python_base_app import log_handling
-
 import atexit
 import errno
 import os
 import signal
 import sys
 import time
+
+from python_base_app import log_handling
+from python_base_app import tools
 
 
 class Daemon(object):
@@ -131,7 +132,7 @@ class Daemon(object):
         #             signal.signal(signal.SIGTERM, sigtermhandler)
         #             signal.signal(signal.SIGINT, sigtermhandler)
 
-        self.log("Daemon gestarted")
+        self.log("Started daemon.")
 
         # Write pidfile
         atexit.register(
@@ -156,7 +157,7 @@ class Daemon(object):
         Start the daemon
         """
 
-        self.log("Daemon starten...")
+        self.log("Starting daemon...")
 
         # Check for a pidfile to see if the daemon already runs
         try:
@@ -183,7 +184,7 @@ class Daemon(object):
         """
 
         if self.verbose >= 1:
-            self.log("Daemon stoppen...")
+            self.log("Stopping daemon...")
 
         # Get the pid from the pidfile
         pid = self.get_pid()
@@ -206,8 +207,10 @@ class Daemon(object):
                 os.kill(pid, signal.SIGTERM)
                 time.sleep(0.1)
                 i = i + 1
-                if i % 10 == 0:
-                    os.kill(pid, signal.SIGHUP)
+
+                if not tools.is_windows():
+                    if i % 10 == 0:
+                        os.kill(pid, signal.SIGHUP)
         except OSError as err:
             if err.errno == errno.ESRCH:
                 if os.path.exists(self.pidfile):
@@ -216,7 +219,7 @@ class Daemon(object):
                 print(str(err))
                 sys.exit(1)
 
-        self.log("Daemon gestoppt")
+        self.log("Daemon stopped")
 
     def restart(self):
         """
